Move more svg files outside of /static
